Rewrite DSL to be more composable
Rewrite the DSL to be more composable. Now the building of the theme
description is based on merging smaller partial descriptions together
instead of mutating a hidden state during evaluation. This way we can
abstract the building of partial descriptions into separate functions as
we see fit.

Further make the building of (internal) theme description and the
conversion of this description into face specifications (that can be fed
to `custom-set-faces` and friend) clearly separate steps. Post-pone
things like converting the colors into their string representation to
the latter step.