Add script for illustration gamut problems
Initially I thought there was a bug in conversion from XYZ space to sRGB
space. It turned out that the bug was caused just by me trying to
convert colors outside the sRGB gamut. So not a bug after all. One can
reproduce them by running the script with `(= L 10.0)`.